+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning].
 ### Added
 
 - Add options and hot-reloading.
+- Add `IServiceCollection.AddCronJob<TCronJob>` extension methods.
 
 ## [1.0.0] - 2023-12-11